+++ src/translations/da/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates da Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Entry GPS Coordinates plugin indlæst',
+    'Search' => 'Søg',
+    'Search Address' => 'Søg adresse',
+    'Coordinates' => 'Koordinater',
+    'Address' => 'Adresse',
+    'Delete Marker' => 'Slet markør',
+];