+++ src/translations/de/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates de Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Entry GPS Coordinates Plugin geladen',
+    'Search' => 'Suchen',
+    'Search Address' => 'Adresse suchen',
+    'Coordinates' => 'Koordinaten',
+    'Address' => 'Adresse',
+    'Delete Marker' => 'Markierung löschen',
+];