+++ src/translations/es/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates es Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Plugin Entry GPS Coordinates cargado',
+    'Search' => 'Buscar',
+    'Search Address' => 'Buscar dirección',
+    'Coordinates' => 'Coordenadas',
+    'Address' => 'Dirección',
+    'Delete Marker' => 'Eliminar marcador',
+];