+++ src/translations/fr/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates fr Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Plugin Entry GPS Coordinates chargé',
+    'Search' => 'Rechercher',
+    'Search Address' => 'Rechercher une adresse',
+    'Coordinates' => 'Coordonnées',
+    'Address' => 'Adresse',
+    'Delete Marker' => 'Supprimer le marqueur',
+];