+++ src/translations/it/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates it Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Plugin Entry GPS Coordinates caricato',
+    'Search' => 'Cerca',
+    'Search Address' => 'Cerca indirizzo',
+    'Coordinates' => 'Coordinate',
+    'Address' => 'Indirizzo',
+    'Delete Marker' => 'Elimina marcatore',
+];