+++ src/translations/nb/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates nb (Norwegian Bokmål) Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Entry GPS Coordinates plugin lastet inn',
+    'Search' => 'Søk',
+    'Search Address' => 'Søk adresse',
+    'Coordinates' => 'Koordinater',
+    'Address' => 'Adresse',
+    'Delete Marker' => 'Slett markør',
+];