+++ src/translations/nl/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates nl Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Entry GPS Coordinates plugin geladen',
+    'Search' => 'Zoeken',
+    'Search Address' => 'Adres zoeken',
+    'Coordinates' => 'Coördinaten',
+    'Address' => 'Adres',
+    'Delete Marker' => 'Markering verwijderen',
+];