+++ src/translations/pl/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates pl Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Wtyczka Entry GPS Coordinates załadowana',
+    'Search' => 'Szukaj',
+    'Search Address' => 'Szukaj adresu',
+    'Coordinates' => 'Współrzędne',
+    'Address' => 'Adres',
+    'Delete Marker' => 'Usuń znacznik',
+];