+++ src/translations/pt/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates pt Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Plugin Entry GPS Coordinates carregado',
+    'Search' => 'Pesquisar',
+    'Search Address' => 'Pesquisar endereço',
+    'Coordinates' => 'Coordenadas',
+    'Address' => 'Endereço',
+    'Delete Marker' => 'Remover marcador',
+];