+++ src/translations/sv/entry-gps-coordinates.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Entry GPS Coordinates plugin for Craft CMS
+ *
+ * Entry GPS Coordinates sv Translation
+ *
+ * @link      https://nthmedia.nl
+ * @copyright Copyright (c) 2019 NTH media
+ */
+
+return [
+    'Entry GPS Coordinates plugin loaded' => 'Entry GPS Coordinates plugin laddat',
+    'Search' => 'Sök',
+    'Search Address' => 'Sök adress',
+    'Coordinates' => 'Koordinater',
+    'Address' => 'Adress',
+    'Delete Marker' => 'Ta bort markör',
+];