+++ src/twigextensions/EntryGpsCoordinatesTwigExtension.php
@@ -45,6 +45,11 @@ class EntryGpsCoordinatesTwigExtension extends \Twig\Extension\AbstractExtension
             new TwigFilter('latitude', [$this, 'latitude']),
             new TwigFilter('longitude', [$this, 'longitude']),
             new TwigFilter('zoomLevel', [$this, 'zoomLevel']),
+            new TwigFilter('streetNumber', [$this, 'streetNumber']),
+            new TwigFilter('route', [$this, 'route']),
+            new TwigFilter('locality', [$this, 'locality']),
+            new TwigFilter('postalCode', [$this, 'postalCode']),
+            new TwigFilter('country', [$this, 'country']),
         ];
     }
 
@@ -60,6 +65,11 @@ class EntryGpsCoordinatesTwigExtension extends \Twig\Extension\AbstractExtension
             new TwigFunction('latitude', [$this, 'latitude']),
             new TwigFunction('longitude', [$this, 'longitude']),
             new TwigFunction('zoomLevel', [$this, 'zoomLevel']),
+            new TwigFunction('streetNumber', [$this, 'streetNumber']),
+            new TwigFunction('route', [$this, 'route']),
+            new TwigFunction('locality', [$this, 'locality']),
+            new TwigFunction('postalCode', [$this, 'postalCode']),
+            new TwigFunction('country', [$this, 'country']),
         ];
     }
 
@@ -116,5 +126,30 @@ class EntryGpsCoordinatesTwigExtension extends \Twig\Extension\AbstractExtension
     public function zoomLevel(?EntryCoordinatesModel $value = null)
     {
         return $value->zoomLevel;
+    }
+
+    public function streetNumber(?EntryCoordinatesModel $value = null)
+    {
+        return $value->streetNumber;
+    }
+
+    public function route(?EntryCoordinatesModel $value = null)
+    {
+        return $value->route;
+    }
+
+    public function locality(?EntryCoordinatesModel $value = null)
+    {
+        return $value->locality;
+    }
+
+    public function postalCode(?EntryCoordinatesModel $value = null)
+    {
+        return $value->postalCode;
+    }
+
+    public function country(?EntryCoordinatesModel $value = null)
+    {
+        return $value->country;
     }
 }